udev: builtin-keyboard: add support for EVDEV_ABS_*
Parse properties in the form
EVDEV_ABS_00="<min>:<max>:<res>:<fuzz>:<flat>"

and apply them to the kernel device. Future processes that open that device
will see the updated EV_ABS range.

This is particularly useful for touchpads that don't provide a resolution in
the kernel driver but can be fixed up through hwdb entries (e.g. bcm5974).

All values in the property are optional, e.g. a string of "::45" is valid to
set the resolution to 45.

The order intentionally orders resolution before fuzz and flat despite it
being the last element in the absinfo struct. The use-case for setting
fuzz/flat is almost non-existent, resolution is probably the most common case
we'll need.

To avoid multiple hwdb invocations for the same device, replace the
hwdb "keyboard:" prefix with "evdev:" and drop the separate 60-keyboard.rules
file. The new 60-evdev.rules is called for all event nodes
anyway, we don't need a separate rules file and second callout to the hwdb
builtin.

(cherry picked from commit 51c0c2869845a058268d54c3111d55d0dd485704)

Changes to the upstream commit:
- 60-keyboard.rules is left in place, the 60-keyboard.hwdb is left
  unmodified. This avoids any potential breakage from user-installed hwdb
  files that are triggered by that rule.

Conflicts:
	hwdb/60-keyboard.hwdb
	rules/60-keyboard.rules

Resolves: #1500119