udev: builtin-keyboard: immediately EVIOCSKEYCODE when we have a pair
Rather than building a map and looping through the map, immediately call the
ioctl when we have a successfully parsed property.

This has a side-effect: before the maximum number of ioctls was limited to the
size of the map (1024), now it is unlimited.

(cherry picked from commit cfba2656e3b4a9c5e03db4ec0a8f76c3762d35a8)

Resolves: #1500119