systemctl: allow disable on the unit file path, but warn about it (#3806)
systemd now returns an error when it is asked to perform disable on the
unit file path. In the past this was allowed, but systemd never really
considered an actual content of the [Install] section of the unit
file. Instead it performed disable on the unit name, i.e. purged all
symlinks pointing to the given unit file (undo of implicit link action
done by systemd when enable is called on the unit file path) and all
symlinks that have the same basename as the given unit file.

However, to notice that [Install] info of the file is not consulted one
must create additional symlinks manually. I argue that in most cases
users do not create such links. Let's be nice to our users and don't
break existing scripts that expect disable to work with the unit file
path.

Fixes #3706.

IMPORTANT
=========
Note that in this downstream backport we actually pass false to
normalize_names(), hence it will not produce any warning when full path
is passed in. This is because we need to preserve behavior compatible
with prior systemd versions shipped in RHEL.

Cherry-picked from: 1d3c86c06fca8311923fcf81af0ab0bbb66e1edd
Resolves: #1348208