load-fragment: fix parsing values in bytes and prevent returning -ERANGE incorrectly
We didn't port our code base to use uint64_t instead of off_t as
upstream did. RLIMIT_INIFINITY is -1ULL and if we cast to off_t (64 bit
signed int on arches we support) then we get -1 and that is always
smaller than correct value returned by parse_size().

To make code changes as minimal as possible (i.e. not port everything
to uint64_t) let's cast off_t to uint64_t and not the other way
around.

RHEL-only

Resolves: #1396277