tests: use fdisk instead of sfdisk
In RHEL7 we have an older version of sfdisk that exits with an error
when executed with sfdisk script that is used in upstream to create
partitions on root disk.

Let's use equivalent fdisk commands to achieve the (more less) same
result.

Also default size of disk image is bumped to 400M. Previous 300M doesn't
work, probably due to some fdisk bug. Size of second partiotion (/var in
TEST-02-CRYPTSETUP) is bumped to 50M to accommodate space requirements
of xfs filesystem.

RHEL-only

Resolves: #1475870