udev: call poll() again after killing the spawned process
Later we check .revents of pfd. Hence we need to initialize it properly
and to do that we need to call poll() once again because previously it
exited twice with timeout and hence left pfd uninitialized. For the
third time it should return immediately since we killed the spawned
process with SIGKILL.

Related: #1697909