avoid possible hang if our child process hangs
If there is one or more unexpected child processes that terminate, but
the "main" child process hangs, we will loop through the terminated
children and then, eventually, get stuck in the waitpid() call. Let's
repeat the main cycle if that situation happens, as that allows us to
finish after timeout.

Related: #1697909